### 2017-12-13 #504 @whedon
[PRE REVIEW]: humanleague: a C++ microsynthesis package with R and python interfaces
**Submitting author:** @virgesmith (<a href="http://orcid.org/0000-0002-9951-6642">Andrew Smith</a>)
**Repository:** <a href="https://github.com/virgesmith/humanleague" target ="_blank">https://github.com/virgesmith/humanleague</a>
**Version:** 2.0.2
