### 2017-12-14 7fa56f4
Changes for CRAN submission #18
