+++ DESCRIPTION
@@ -1,10 +1,15 @@
 Package: humanleague
 Type: Package
 Title: Synthetic Population Generator
-Description: Generates high-entropy integer synthetic populations using quasirandom sampling, in arbitrary dimensionality. Also implements IPF for reference.
 Version: 2.0.2
-Author: Andrew Smith
-Maintainer: Andrew Smith <a.p.smith@leeds.ac.uk>
+Description: Generates high-entropy integer synthetic populations from marginal and (optionally) seed data using quasirandom sampling, 
+  in arbitrary dimensionality (Smith, Lovelace and Birkin (2017) <doi:10.18564/jasss.3550>). 
+  The package also provides an implementation of the iterative Iterative Proportional Fitting (IPF) algorithm (Zaloznik (2011) <doi:10.13140/2.1.2480.9923>).
+Authors@R: c(
+  person("Andrew", "Smith", email="a.p.smith@leeds.ac.uk", role=c("aut", "cre")),
+  person("Steven", "Johnson", role=c("ctb","cph"), comment="Sobol sequence generator implementation"),
+  person("John", "Burkhardt", role=c("ctb","cph"), comment="C++ implementation of incomplete gamma function"),
+  person("G", "Bhattacharjee", role=c("ctb"), comment="Original FORTRAN implementation of incomplete gamma function"))
 License: MIT + file LICENCE
 Encoding: UTF-8
 LazyData: true
