### 2017-12-14 0e50df6
More CRAN fixes #18
