### 2017-12-18 #504 @leeper
@jordigh would you be interested in reviewing this (or at least the C++ and Python components) for JOSS?
