### 2018-01-04 #504 @leeper
@hausen Would you be interested in reviewing this for JOSS?
