### 2018-01-05 #504 @leeper
@wrathematics Would you be interested in reviewing this for JOSS?
