### 2018-01-10 #504 @wrathematics
I'm catching up from the holidays + a nasty cold, so I have some other things to attend to right now. If you don't find anyone else, I believe I can review it in a week or two.
