### 2018-03-04 #504 @leeper
@wrathematics Would you still be interested?
