### 2018-03-10 #504 @leeper
@hausen Would you be interested in reviewing this for JOSS?
