### 2018-03-14 #504 @virgesmith
sadly not. I can't think of anyone with both C++ skills and the domain expertise
