+++ .travis.yml
@@ -38,7 +38,8 @@ matrix:
       language: generic
       before_install:
         - brew update
-        - brew install python3
+        - brew upgrade python
+        #- brew install python3
         - pip3 install numpy distutils_pytest pytest
       script:
         - python3 setup.py install
