### 2018-03-21 46f5c1b
allow zero population in probToIntFreq #20
