+++ humanleague/py_api.cpp
@@ -37,16 +37,16 @@ pycpp::List flatten(const size_t pop, const NDArray<T>& t)
 // flatten n-D integer array into 2-d table
 extern "C" PyObject* humanleague_flatten(PyObject* self, PyObject* args)
 {
-  try 
+  try
   {
     PyObject* arrayArg;
 
     // args e.g. "s" for string "i" for integer, "d" for float "ss" for 2 strings
     if (!PyArg_ParseTuple(args, "O!", &PyArray_Type, &arrayArg))
-      return nullptr; 
+      return nullptr;
 
     pycpp::Array<int64_t> pyarray(arrayArg);
-    
+
     NDArray<int64_t> array(pyarray.toNDArray());
 
     size_t pop = 0;
@@ -61,7 +61,7 @@ extern "C" PyObject* humanleague_flatten(PyObject* self, PyObject* args)
     {
       outer.set(i, pycpp::List(list[i]));
     }
- 
+
     return outer.release();
   }
   catch(const std::exception& e)
@@ -76,7 +76,7 @@ extern "C" PyObject* humanleague_flatten(PyObject* self, PyObject* args)
 
 extern "C" PyObject* humanleague_prob2IntFreq(PyObject* self, PyObject* args)
 {
-  try 
+  try
   {
     PyObject* probArg;
     int pop;
@@ -84,14 +84,14 @@ extern "C" PyObject* humanleague_prob2IntFreq(PyObject* self, PyObject* args)
     // args e.g. "s" for string "i" for integer, "d" for float "ss" for 2 strings
     if (!PyArg_ParseTuple(args, "O!i", &PyArray_Type, &probArg, &pop))
       return nullptr;
-      
+
     const std::vector<double>& prob = pycpp::Array<double>(probArg).toVector<double>();
 
     double var;
 
-    if (pop < 1)
+    if (pop < 0)
     {
-      throw std::runtime_error("population must be strictly positive");
+      throw std::runtime_error("population cannot be negative");
     }
 
     if (std::fabs(std::accumulate(prob.begin(), prob.end(), -1.0)) > 1000*std::numeric_limits<double>::epsilon())
@@ -119,7 +119,7 @@ extern "C" PyObject* humanleague_prob2IntFreq(PyObject* self, PyObject* args)
 // prevents name mangling (but works without this)
 extern "C" PyObject* humanleague_sobol(PyObject *self, PyObject *args)
 {
-  try 
+  try
   {
     int dim, length, skips = 0;
 
@@ -137,7 +137,7 @@ extern "C" PyObject* humanleague_sobol(PyObject *self, PyObject *args)
     {
       sequence[idx] = sobol() * scale;
     }
-    
+
     pycpp::Array<double> result(sequence);
 
     return result.release();
@@ -155,16 +155,16 @@ extern "C" PyObject* humanleague_sobol(PyObject *self, PyObject *args)
 // prevents name mangling (but works without this)
 extern "C" PyObject* humanleague_ipf(PyObject *self, PyObject *args)
 {
-  try 
+  try
   {
     PyObject* indexArg;
     PyObject* arrayArg;
     PyObject* seedArg;
-    
+
     // args e.g. "s" for string "i" for integer, "d" for float "ss" for 2 strings
     if (!PyArg_ParseTuple(args, "O!O!O!", &PyArray_Type, & seedArg, &PyList_Type, &indexArg, &PyList_Type, &arrayArg))
       return nullptr;
-    
+
     // seed
     pycpp::Array<double> seed(seedArg);
     // expects a list of numpy arrays containing int64
@@ -178,8 +178,8 @@ extern "C" PyObject* humanleague_ipf(PyObject *self, PyObject *args)
[…]
