### 2018-04-24 #629 @rdrr1990
Apologies again for delay! New to JOSS--do I just post things as they come up? On `R` side, just built package, `testthat` runs fine but `R CMD check --as-cran` returns the classic nuisance error:

```
Found the following (possibly) invalid URLs:
  URL: http://cran.r-project.org/package=humanleague
    From: README.md
    CRAN URL not in canonical form
```
