### 2018-04-24 #629 @leeper
@rdrr1990 Easiest thing to do is open issues on the source repo: https://github.com/virgesmith/humanleague. If you want to provide a summary response and/or say anything that doesn't make sense as a standalone issue, feel free to add it here.
