### 2018-04-29 #629 @rdrr1990
`humanleague` offers users the ability to quickly microsynthesize populations from marginals and (optionally) seed data. The package installs nicely for either `R` or `Python` and the extensive unit tests run without a hitch on both platforms. The software also builds on a companion paper, [Population Synthesis with Quasirandom Integer Sampling](http://jasss.soc.surrey.ac.uk/20/4/14.html) which explains the motivation behind Quasirandom Integer Sampling (QIWPS where the 'W' denotes without replacement) as opposed to pseudorandom Iterative Proportional Fitting (IPF). Quasirandom numbers fill the sample space evenly and do not exhibit clumping and other little irregularities like pseudorandom numbers, which the authors argue can be preferable for at least certain applications. The big claim is that QIWPS does a better job at maximizing entropy than IPF.

The paper documents considerable runtime decreases for QIWPS vs. IPF (though the authors acknowledge it's is not an apples-to-apples comparison because it is `R` vs `C++`). The paper also demonstrates the R syntax, which is very straightforward. The error messages are also perfectly readable if you e.g. attempt to provide marginals that do not sum to one, which is also appreciated by users.

Overall, `humanleague` makes a very nice contribution to open source statistical software. My one comment (intended as a recommendation, not a criticism) is that the authors consider writing a vignette (or Github gist, etc.) that takes a practical problem to introduce users to the package. Though the paper is clear enough, most practitioners won't be looking for a discussion of the difference between pseudo and quasi random numbers when they attempt to learn how/whether to use the library. The authors mention ABM as a promising application, so that could certainly be useful. Or pick a recent survey and start with the marginals. Since most will be familiar with $\chi^2$  statistic like the one used in the simulation, that could be an easy way to introduce people to the advantages of `humanleague`.
