### 2018-04-30 #629 @rdrr1990
Just checked them off... In checking the boxes, I realize I forgot to mention that the authors should add [distutils-pytest](https://pypi.org/project/distutils-pytest/) to the readme's Python dependencies. (Along with the stated `numpy` that's all I needed to build...). Otherwise happy to give my recommendation for publication. Cheers,
