### 2018-04-30 2d6b3a0
fixes #21
