+++ R/RcppExports.R
@@ -46,9 +46,9 @@ synthPopG <- function(marginals, exoProbsIn) {
 #' }
 #' @examples
 #' ageByGender = array(c(1,2,5,3,4,3,4,5,1,2), dim=c(5,2))
-#' ageByEthnicity = array(c(4,6,5,6,4,5), dim=c(3,2))
+#' ethnicityByGender = array(c(4,6,5,6,4,5), dim=c(3,2))
 #' seed = array(rep(1,30), dim=c(5,2,3))
-#' result = ipf(seed, list(c(1,2), c(3,2)), list(ageByGender, ageByEthnicity))
+#' result = ipf(seed, list(c(1,2), c(3,2)), list(ageByGender, ethnicityByGender))
 #' @export
 ipf <- function(seed, indices, marginals) {
     .Call('_humanleague_ipf', PACKAGE = 'humanleague', seed, indices, marginals)
@@ -70,8 +70,8 @@ ipf <- function(seed, indices, marginals) {
 #' }
 #' @examples
 #' ageByGender = array(c(1,2,5,3,4,3,4,5,1,2), dim=c(5,2))
-#' ageByEthnicity = array(c(4,6,5,6,4,5), dim=c(3,2))
-#' result = qis(list(c(1,2), c(3,2)), list(ageByGender, ageByEthnicity))
+#' ethnicityByGender = array(c(4,6,5,6,4,5), dim=c(3,2))
+#' result = qis(list(c(1,2), c(3,2)), list(ageByGender, ethnicityByGender))
 #' @export
 qis <- function(indices, marginals, skips = 0L) {
     .Call('_humanleague_qis', PACKAGE = 'humanleague', indices, marginals, skips)
@@ -94,9 +94,9 @@ qis <- function(indices, marginals, skips = 0L) {
 #' }
 #' @examples
 #' ageByGender = array(c(1,2,5,3,4,3,4,5,1,2), dim=c(5,2))
-#' ageByEthnicity = array(c(4,6,5,6,4,5), dim=c(3,2))
+#' ethnicityByGender = array(c(4,6,5,6,4,5), dim=c(3,2))
 #' seed = array(rep(1,30), dim=c(5,2,3))
-#' result = qisi(seed, list(c(1,2), c(3,2)), list(ageByGender, ageByEthnicity))
+#' result = qisi(seed, list(c(1,2), c(3,2)), list(ageByGender, ethnicityByGender))
 #' @export
 qisi <- function(seed, indices, marginals, skips = 0L) {
     .Call('_humanleague_qisi', PACKAGE = 'humanleague', seed, indices, marginals, skips)
