+++ requirements.txt
@@ -1,3 +1,4 @@
+numpy
 distutils_pytest
 pytest
 
