+++ man/qisi.Rd
@@ -30,7 +30,7 @@ C++ QIS-IPF implementation
 }
 \examples{
 ageByGender = array(c(1,2,5,3,4,3,4,5,1,2), dim=c(5,2))
-ageByEthnicity = array(c(4,6,5,6,4,5), dim=c(3,2))
+ethnicityByGender = array(c(4,6,5,6,4,5), dim=c(3,2))
 seed = array(rep(1,30), dim=c(5,2,3))
-result = qisi(seed, list(c(1,2), c(3,2)), list(ageByGender, ageByEthnicity))
+result = qisi(seed, list(c(1,2), c(3,2)), list(ageByGender, ethnicityByGender))
 }
