### 2018-04-30 #629 @virgesmith
I've corrected requirements.txt (its numpy that's missing, not distutils-pytest), I've also corrected the documentation that a user recently pointed out was misleading (i.e. wrong), I've bumped the version and will create a new release unless theres anything else
