### 2018-04-30 #629 @rdrr1990
Oh, missed that .txt. I just meant in https://github.com/virgesmith/humanleague/blob/master/README.md "Requires Python 3, with numpy, pytest, and distutils-pytest installed" would be clearer (ofc not a big deal). Good to go as far as I'm concerned, great work on the package!
