### 2018-04-30 3e99649
Update README.md
mention test package deps for python install #21
