### 2018-05-01 #629 @leeper
@rdrr1990 Excellent. Thank you, again!

@virgesmith Please update your Zenodo archive and post the new DOI here. We're nearly there!
