### 2018-05-01 #629 @leeper
Okay, everything looks good. Congrats, @virgesmith, your paper is now accepted at JOSS! (Sorry for the longer-than-expected review process.)

@rdrr1990 Thank you again for your willingness to review for us!

@arfon Over to you.
